refactor(core): extract storage tests
